Fix deleterevisions user choice prompt to work with Python 2
When prompting the user for a y/n choice, the deleterevisions command
was failing for me in Python 2.7.9 because it was falling back to use
input() instead of raw_input().

This fallback is appropriate for Python 3 in which raw_input() was
replaced by input(), but should never be done in Python 2 because there
the input() function evaluates the user input as code.

I am not really sure why the attempted Python 2/3 recognition technique
of assigning the raw_input function to itself fails to work correctly
for me, but this restructured version should accomplish the same thing
with less chance of incorrectly assuming it is in Python 3.