Add --no-input so management commands look similar
When Django management commands have a means of skipping confirmation
checks, usually the flag is '--no-input'. The 'deleterevisons' command
uses '--no-confirmation'. Since it functions exactly as you would expect
'--no-input', it makes sense to follow convention. The existing
'--no-confirmation' flag is left deprecated, but there is no schedule
for removal, nor is there a need to remove it any time soon.